feat: LanguageGameView 대치
