+++ SoloDeveloperTraining/SoloDeveloperTraining/Production/Presentation/Work/Game/LanguageGameView.swift
@@ -7,61 +7,34 @@
 
 import SwiftUI
 
-private enum Constant {
-    enum Padding {
-        static let horizontal: CGFloat = 16
-        static let toolBarBottom: CGFloat = 10
-    }
-
-    enum Spacing {
-        static let itemHorizontal: CGFloat = 25
-        static let buttonHorizontal: CGFloat = 17
-    }
-
-    enum Game {
-        static let itemCount: Int = 5
-        static let feverDecreaseInterval: Double = 0.1
-        static let feverDecreasePercentPerTick: Double = 5
-    }
-
-    enum EffectLabel {
-        static let offsetY: CGFloat = -34
-    }
-}
+import DUDesignSystem
 
 struct LanguageGameView: View {
-    // MARK: Properties
-    let user: User
 
-    /// 게임에 사용되는 언어 타입 목록
-    private let languageTypeList: [LanguageType] = [
-        .swift,
-        .kotlin,
-        .dart,
-        .python
-    ]
-
-    // MARK: State Properties
-    /// 게임 시작 상태 (부모 뷰와 바인딩)
-    @Binding var isGameStarted: Bool
-    @Binding var gameActionGoldDelta: Int
-    @Binding var tabSwitchPause: Bool
-
-    /// 상태를 유지
+    /// 언어 맞추기 게임 모델
     @State private var game: LanguageGame
+    /// 닫기 버튼으로 인한 일시정지
     @State private var closePause: Bool = false
-
-    /// 획득한 골드를 표시하기 위한 효과 라벨 배열
-    @State private var effectValues: [(id: UUID, value: Int)] = []
-
-    /// 현재 진행 중인 언어 버튼 탭 Task
+    /// 진행 중인 언어 버튼 탭 Task
     @State private var currentActionTask: Task<Void, Never>?
+    /// 획득한 골드를 표시하는 효과 라벨 목록
+    @State private var effectLabels: [EffectLabelData] = []
 
-    // 광고 팝업 관련
+    /// 게임 시작 여부 (false로 바꾸면 선택 화면으로 복귀)
+    @Binding var isGameStarted: Bool
+    /// 이번 세션에서 획득한 골드 누적량
+    @Binding var gameActionGoldDelta: Int
+    /// 탭 전환으로 인한 일시정지
+    @Binding var tabSwitchPause: Bool
+    /// 광고 시청 후 음료 지급 팝업 표시 여부
     @Binding var showDrinkAdPopup: Bool
+    /// 나가기 보너스 팝업 표시 여부
     @Binding var showExitBonusPopup: Bool
+    /// 광고 팝업에서 선택된 음료 타입
     @Binding var selectedDrinkType: ConsumableType?
+    /// 팝업에서 게임 재개 시 호출되는 콜백
     @Binding var resumeGameCallback: (() -> Void)?
+    /// 팝업에서 게임 종료 시 호출되는 콜백
     @Binding var exitGameCallback: (() -> Void)?
 
     init(
@@ -69,84 +42,138 @@ struct LanguageGameView: View {
         isGameStarted: Binding<Bool>,
         gameActionGoldDelta: Binding<Int>,
         tabSwitchPause: Binding<Bool>,
-        animationSystem: CharacterAnimationSystem? = nil,
+        animationSystem: CharacterAnimationSystem?,
         showDrinkAdPopup: Binding<Bool>,
         showExitBonusPopup: Binding<Bool>,
         selectedDrinkType: Binding<ConsumableType?>,
         resumeGameCallback: Binding<(() -> Void)?>,
         exitGameCallback: Binding<(() -> Void)?>
     ) {
-        self._isGameStarted = isGameStarted
-        self._gameActionGoldDelta = gameActionGoldDelta
-        self._tabSwitchPause = tabSwitchPause
-        self.user = user
-        self._showDrinkAdPopup = showDrinkAdPopup
-        self._showExitBonusPopup = showExitBonusPopup
-        self._selectedDrinkType = selectedDrinkType
-        self._resumeGameCallback = resumeGameCallback
-        self._exitGameCallback = exitGameCallback
-
-        // 게임 초기화
         let game = LanguageGame(
             user: user,
             feverSystem: .init(
-                decreaseInterval: Constant.Game.feverDecreaseInterval,
-                decreasePercentPerTick: Constant.Game.feverDecreasePercentPerTick
+                decreaseInterval: 0.1,
+                decreasePercentPerTick: 5
             ),
             buffSystem: .init(),
-            itemCount: Constant.Game.itemCount,
+            itemCount: 5,
             animationSystem: animationSystem
         )
-        self._game = State(initialValue: game)
-        self.game.startGame()
+        game.startGame()
+        _game = State(initialValue: game)
+        _isGameStarted = isGameStarted
+        _gameActionGoldDelta = gameActionGoldDelta
+        _tabSwitchPause = tabSwitchPause
+        _showDrinkAdPopup = showDrinkAdPopup
+        _showExitBonusPopup = showExitBonusPopup
+        _selectedDrinkType = selectedDrinkType
+        _resumeGameCallback = resumeGameCallback
+        _exitGameCallback = exitGameCallback
     }
 
     var body: some View {
-        GeometryReader { geometry in
-            VStack(alignment: .center, spacing: 0) {
-                // 상단 툴바 (닫기, 아이템 버튼, 피버 게이지)
+        GeometryReader { _ in
+            VStack(spacing: 0) {
                 toolbarSection
-                Spacer()
-                // 중앙 언어 아이템 영역 (획득 골드 효과 포함)
-                languageItemsSection
-                Spacer()
-                // 하단 언어 선택 버튼 영역
-                languageButtonsSection
-                Spacer()
+                gameAreaSection
             }
             .onAppear {
-                // 게임 재개 콜백 설정
                 resumeGameCallback = { [weak game] in
                     game?.resumeGame()
                 }
                 exitGameCallback = { handleCloseButton() }
             }
-            .pauseGameStyle(
-                pauseBinding: pauseBinding,
-                height: geometry.size.height,
-                onLeave: { showExitBonusPopup = true },
-                onPause: { game.pauseGame() },
-                onResume: { game.resumeGame() }
-            )
         }
     }
 }
 
-// MARK: - View Components
+// MARK: - Sections
 private extension LanguageGameView {
-    /// 상단 툴바
+
     var toolbarSection: some View {
-        OldGameToolBar(
-            closeButtonDidTapHandler: { closePause = true },
-            coffeeButtonDidTapHandler: { useConsumableItem(.coffee) },
-            energyDrinkButtonDidTapHandler: { useConsumableItem(.energyDrink) },
-            feverState: game.feverSystem,
-            buffSystem: game.buffSystem,
-            coffeeCount: .constant(game.user.inventory.count(.coffee) ?? 0),
-            energyDrinkCount: .constant(game.user.inventory.count(.energyDrink) ?? 0)
+        GameToolBar(
+            feverStage: game.feverSystem.feverStage,
+            feverProgress: {
+                let stageBase = Double(game.feverSystem.feverStage) * 100.0
+                return (game.feverSystem.feverPercent - stageBase) / 100.0
+            }(),
+            feverMultiplier: game.feverSystem.feverStage == 0 ? 0 : game.feverSystem.feverMultiplier,
+            coffeeCount: game.user.inventory.count(.coffee) ?? 0,
+            energyDrinkCount: game.user.inventory.count(.energyDrink) ?? 0,
+            coffeeCooldown: {
+                Double(game.buffSystem.coffeeDuration) / Double(ConsumableType.coffee.duration)
+            }(),
+            energyDrinkCooldown: {
+                Double(game.buffSystem.energyDrinkDuration) / Double(ConsumableType.energyDrink.duration)
+            }(),
+            onClose: {
+                closePause = true
+                SoundService.shared.stopAllSFX()
+                SoundService.shared.trigger(.buttonTap)
+            },
+            onCoffee: { useConsumableItem(.coffee) },
+            onEnergyDrink: { useConsumableItem(.energyDrink) }
         )
-        .padding(.horizontal, Constant.Padding.horizontal)
-        .padding(.bottom, Constant.Padding.toolBarBottom)
+        .padding(.bottom, TokenSpacing.md)
+    }
+
+    var gameAreaSection: some View {
+        VStack(spacing: 0) {
+            languageItemsSection
+            languageButtonsSection
+        }
+        .gamePauseWrapper(
+            pauseBinding: pauseBinding,
+            onLeave: { showExitBonusPopup = true },
+            onPause: { game.pauseGame() },
+            onResume: { game.resumeGame() }
+        )
+    }
+
+    var languageItemsSection: some View {
+        HStack(spacing: TokenSpacing.lg) {
+            ForEach(Array(game.itemList.enumerated()), id: \.offset) { _, item in
+                if item.languageType == .empty {
+                    Color.clear.frame(
+                        width: TokenIconSize.size38.rawValue,
+                        height: TokenIconSize.size38.rawValue
+                    )
+                } else {
+                    LanguageItem(
+                        language: mapLanguageType(item.languageType),
+                        state: mapLanguageItemState(item.state)
+                    )
+                }
+            }
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .overlay(alignment: .top) {
+            ZStack {
+                ForEach(effectLabels) { data in
+                    EffectLabel(type: data.value >= 0 ? .plus : .minus, text: "\(abs(data.value))") {
+                        removeEffectLabel(id: data.id)
+                    }
+                }
+            }
+        }
+    }
+
+    var languageButtonsSection: some View {
+        HStack(spacing: TokenSpacing.lg) {
+            LanguageItemButton(language: .swift) { handleLanguageButtonTap(.swift) }
+            LanguageItemButton(language: .kotlin) { handleLanguageButtonTap(.kotlin) }
+            LanguageItemButton(language: .dart) { handleLanguageButtonTap(.dart) }
+            LanguageItemButton(language: .python) { handleLanguageButtonTap(.python) }
+        }
+        .frame(maxWidth: .infinity)
+        .padding(.top, TokenSpacing.md)
+        .padding(.bottom, TokenGrid.paddingBottom)
+        .frame(height: 130)
+        .overlay(alignment: .top) {
+            Rectangle()
+                .fill(Color.pastelBlueGray)
+                .frame(height: 1)
+        }
     }
 
     var pauseBinding: Binding<Bool> {
@@ -158,64 +185,45 @@ private extension LanguageGameView {
             }
         )
     }
-
-    /// 중앙 언어 아이템 영역
-    var languageItemsSection: some View {
-        HStack(alignment: .bottom, spacing: Constant.Spacing.itemHorizontal) {
-            ForEach(Array(game.itemList.enumerated()), id: \.offset) { _, item in
-                OldLanguageItem(
-                    languageType: item.languageType,
-                    state: item.state
-                )
-            }
-        }
-        .frame(maxWidth: .infinity)
-        .overlay(alignment: .top) {
-            // 획득한 골드를 표시하는 효과 라벨
-            ZStack {
-                ForEach(effectValues, id: \.id) { effect in
-                    OldEffectLabel(value: effect.value) {
-                        removeEffectLabel(id: effect.id)
-                    }
-                }
-            }
-            .offset(y: Constant.EffectLabel.offsetY)
-        }
-    }
-
-    /// 하단 언어 선택 버튼 영역
-    var languageButtonsSection: some View {
-        HStack(spacing: Constant.Spacing.buttonHorizontal) {
-            ForEach(languageTypeList, id: \.self) { type in
-                LanguageButton(languageType: type) {
-                    handleLanguageButtonTap(type)
-                }
-            }
-        }
-    }
 }
 
-// MARK: - Actions
+// MARK: - Helper
 private extension LanguageGameView {
-    /// 닫기 버튼 클릭 처리
-    func handleCloseButton() {
-        // 진행 중인 액션 Task 취소
-        currentActionTask?.cancel()
-        currentActionTask = nil
 
-        game.stopGame()
-        isGameStarted = false
+    func mapLanguageType(_ type: LanguageType) -> LanguageItem.LanguageType {
+        switch type {
+        case .swift:  return .swift
+        case .kotlin: return .kotlin
+        case .dart:   return .dart
+        case .python: return .python
+        case .empty:  return .swift // .empty는 뷰에서 Color.clear로 처리
+        }
     }
 
-    /// 언어 버튼 클릭 처리
-    func handleLanguageButtonTap(_ type: LanguageType) {
-        // 이전 액션이 진행 중이면 취소
+    func mapLanguageItemState(_ state: LanguageItemState) -> LanguageItem.LanguageItemState {
+        switch state {
+        case .completed: return .completed
+        case .active:    return .active
+        case .upcoming:  return .upcoming
+        case .empty:     return .upcoming // .empty는 뷰에서 Color.clear로 처리
+        }
+    }
+
+    func mapToAppLanguageType(_ type: LanguageItem.LanguageType) -> LanguageType {
+        switch type {
+        case .swift:  return .swift
+        case .kotlin: return .kotlin
+        case .dart:   return .dart
+        case .python: return .python
+        }
+    }
+
+    func handleLanguageButtonTap(_ type: LanguageItem.LanguageType) {
         currentActionTask?.cancel()
 
         currentActionTask = Task {
-            let gainedGold = await game.didPerformAction(type)
+            let gainedGold = await game.didPerformAction(mapToAppLanguageType(type))
 
-            // Task가 취소되었으면 UI 업데이트 생략
             guard !Task.isCancelled else { return }
 
             SoundService.shared.trigger(gainedGold > 0 ? .languageCorrect : .languageWrong)
@@ -223,16 +231,29 @@ private extension LanguageGameView {
                 HapticService.shared.trigger(.error)
             }
             gameActionGoldDelta += gainedGold
-            showEffectLabel(gainedGold: gainedGold)
+            showEffectLabel(value: gainedGold)
         }
     }
 
-    /// 소비 아이템 사용 처리
+    func showEffectLabel(value: Int) {
+        let data = EffectLabelData(id: UUID(), position: .zero, value: value)
+        effectLabels.append(data)
+    }
+
+    func removeEffectLabel(id: UUID) {
+        effectLabels.removeAll { $0.id == id }
+    }
+
+    func handleCloseButton() {
+        currentActionTask?.cancel()
+        currentActionTask = nil
+        game.stopGame()
+        isGameStarted = false
+    }
+
     func useConsumableItem(_ type: ConsumableType) {
         let count = game.user.inventory.count(type) ?? 0
-
         if count > 0 {
-            // 음료 사용
             if game.user.inventory.drink(type) {
                 SoundService.shared.trigger(.itemConsume)
                 HapticService.shared.trigger(.success)
@@ -240,60 +261,9 @@ private extension LanguageGameView {
                 game.user.record.record(type == .coffee ? .coffeeUse : .energyDrinkUse)
             }
         } else {
-            // 광고 팝업 표시
             selectedDrinkType = type
             showDrinkAdPopup = true
             game.pauseGame()
         }
     }
-}
-
-// MARK: - Helper Methods
-private extension LanguageGameView {
-    /// 획득한 골드를 표시하는 효과 라벨 추가
-    /// - Parameter gainedGold: 획득한 골드 (음수일 경우 손실)
-    func showEffectLabel(gainedGold: Int) {
-        let effectId = UUID()
-        effectValues.append((id: effectId, value: gainedGold))
-    }
-
-    /// 효과 라벨 제거 (애니메이션 완료 시 콜백으로 호출)
-    /// - Parameter id: 제거할 효과 라벨의 ID
-    func removeEffectLabel(id: UUID) {
-        effectValues.removeAll { $0.id == id }
-    }
-}
-
-#Preview {
-    @Previewable @State var isGameStarted = true
-    @Previewable @State var gameActionGoldDelta = 0
-    @Previewable @State var tabSwitchPause = true
-    @Previewable @State var showDrinkAdPopup = false
-    @Previewable @State var showExitBonusPopup = false
-    @Previewable @State var selectedDrinkType: ConsumableType?
-    @Previewable @State var resumeGameCallback: (() -> Void)?
-    @Previewable @State var exitGameCallback: (() -> Void)?
-
-    let user = User(
-        nickname: "Test",
-        wallet: .init(),
-        inventory: .init(),
-        record: .init(),
-        skills: [
-            .init(key: SkillKey(game: .language, tier: .beginner), level: 1000)
-        ]
-    )
-
-    LanguageGameView(
-        user: user,
-        isGameStarted: $isGameStarted,
-        gameActionGoldDelta: $gameActionGoldDelta,
-        tabSwitchPause: $tabSwitchPause,
-        animationSystem: nil,
-        showDrinkAdPopup: $showDrinkAdPopup,
-        showExitBonusPopup: $showExitBonusPopup,
-        selectedDrinkType: $selectedDrinkType,
-        resumeGameCallback: $resumeGameCallback,
-        exitGameCallback: $exitGameCallback
-    )
 }
